requirements.txt
pyserial==2.7
tweepy==2.1
